allow for complex dipole moments
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@4981 6419b465-d902-0410-997c-be2266baf723

+++ gpaw/grid_descriptor.py
@@ -525,7 +525,7 @@ class GridDescriptor(Domain):
         rho_cg = [np.sum(rho_xy, axis=1),
                   np.sum(rho_xy, axis=0),
                   np.sum(rho_xz, axis=0)]
-        d_c = np.zeros(3)
+        d_c = np.zeros(3, dtype=rho_xyz.dtype)
         for c in range(3):
             r_g = (np.arange(self.n_c[c], dtype=float) +
                    self.beg_c[c]) * self.h_c[c]
